Activate logging by default #31
I don't see why logging shouldn't be the default. Non-developers will appreciate it to see what is going on in the programme, and so will developers in doubt when debugging. If devs want to deactivate logging they can set the parameter to `No`. This appears more functional than inputting this information or writing a custom workaround to avoid user-input with each execution.